Reorganize Framework: move polyfills into Polyfills/System.Text/
Encoding span overloads, StringBuilder.ChunkEnumerator, StringBuilder.Append(ROS<char>) extensions move to namespace System.Text. char.IsAsciiLetter polyfill moves to namespace System under Polyfills/System/CharExtensions.cs.
