+++ touki/Framework/Polyfills/System.Text/ChunkEnumerator.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) 2025 Jeremy W Kuhne
+// Copyright (c) 2025 Jeremy W Kuhne
 // SPDX-License-Identifier: MIT
 // See LICENSE file in the project root for full license information
 
@@ -9,7 +9,9 @@
 
 using System.ComponentModel;
 
-namespace Touki.Text;
+using Touki.Text;
+
+namespace System.Text;
 
 /// <summary>
 ///  ChunkEnumerator supports both the IEnumerable and IEnumerator pattern so foreach
