+++ touki/Framework/Polyfills/System.Text/EncodingExtensions.cs
