+++ touki/Framework/Polyfills/System.Text/StringBuilderExtensions.cs
@@ -1,8 +1,10 @@
-﻿// Copyright (c) 2025 Jeremy W Kuhne
+// Copyright (c) 2025 Jeremy W Kuhne
 // SPDX-License-Identifier: MIT
 // See LICENSE file in the project root for full license information
 
-namespace Touki.Text;
+using Touki;
+
+namespace System.Text;
 
 /// <summary>
 ///  Extension methods for <see cref="StringBuilder"/>.
